Refactor the <v-grid>-component
