--- resources/frontend/src/app/components/grid/grid.html
@@ -1,5 +1,0 @@
-<div :class="classNames">
-  <slot>
-    {{ content }}
-  </slot>
-</div>
