--- resources/frontend/src/app/components/grid/grid.styl
@@ -1,12 +1,0 @@
-@import './../../../assets/stylus/imports.styl'
-
-.grid
-  padding: 0 .5rem
-
-  +variant(container)
-    margin: 0 auto
-    max-width: 1200px
-    width: 100%
-
-  +variant(no-gutters)
-    padding: 0
