+++ resources/frontend/src/app/components/grid/index.js
@@ -1,0 +1,7 @@
+import * as Grid from './grid';
+
+export const VGrid = Grid;
+
+export default {
+  VGrid,
+};
