+++ resources/frontend/src/app/pages/not-found/index/index.js
@@ -7,7 +7,7 @@
 
 import * as VLayout from './../../../layouts/base/base.vue';
 import { VCard, VCardBody } from './../../../components/card';
-import * as VGrid from './../../../components/grid/grid.vue';
+import { VGrid } from './../../../components/grid';
 import * as VRow from './../../../components/row/row.vue';
 import { VCol } from './../../../components/col';
