+++ resources/frontend/src/app/pages/scammers/show/show.js
@@ -9,7 +9,7 @@ import Proxy from './../../../proxies/ScammerProxy';
 import Transformer from './../../../transformers/ScammerTransformer';
 import * as VLayout from './../../../layouts/base/base.vue';
 import { VCard, VCardBody } from './../../../components/card';
-import * as VGrid from './../../../components/grid/grid.vue';
+import { VGrid } from './../../../components/grid';
 import * as VRow from './../../../components/row/row.vue';
 import { VCol } from './../../../components/col';
